Pin integer widths in ITensors Serialized* layouts
Cross-language readability changes to the on-disk Serialized* layouts:
- version :: UInt32 (matches Base.VersionNumber's field width).
- QNVal.val, modulus, QNSpace.dims, Index.plev :: pinned to Int64.
- Index.dir :: Int8, documented as In=-1, Neither=0, Out=+1.
- Index.id :: stays UInt64.

+++ ext/ITensorsJLD2Ext/index.jl
@@ -5,9 +5,9 @@ using JLD2: JLD2
 JLD2.writeas(::Type{<:Index{S}}) where {S} = SerializedIndex{_writeas_space(S)}
 
 function JLD2.wconvert(::Type{<:SerializedIndex}, i::Index)
-    version = 1
+    version = UInt32(1)
     return SerializedIndex(
-        version, id(i), _wconvert_space(space(i)), Int(dir(i)),
+        version, id(i), _wconvert_space(space(i)), Int8(dir(i)),
         JLD2.wconvert(SerializedTagSet, tags(i)), plev(i)
     )
 end
